More WIP
When a consumer crashes, we now remove the pids it had taken from the
pool via fail.

Also started adding some stats which might provide for a cleaner way to
do some of the testing; take a pid and then inspect the stats, induce a
crash, and then inspect stats again.